add travis ci configuration
Add a configuration for Travis CI that will run flake8 and unit tests
on various versions of Python.

I've ignored E501 from flake8, which is that a certain line exceeds 80
characters. Many lines violate this, and I think fixing it in many
cases harms readability, so I'm inclined to just keep an exclusion for
this rule.