Fix up places that treated warnings as Exceptions and enable doctests (#7)
* Fix documentation: warnings aren't exceptions

run() will warn, not raise an exception, if the FPGA is already
running. Fix examples to reflect this.

Also fixup the indentation.

* remove dead code

Since warnings aren't exceptions, this code path is dead.

* Fix documentation and turn it into a doctest

Rework the documentation to demonstrate how to handle a warning, and
turn it into a doctest so it doesn't become stale again.