Removing string based implementation
Now both read and write will calculate the read and write values using
the calculated delta.  This as a response to the review. In doing so I
was able to remove the fixedPointHelper, the only thing left in that
file was the warn, which I moved to a method of session._FXPRegister

I updated the tests to now work with the correct implementation my next
commit will include a test clean up