Fxp support (#15)
* Starting to lay down some work

* First pass at FXP no testing yet

* Fixing bugs and adding unit tests

* Adding more unit tests

* Have bitfile parsing cast correctly so FXP_Register contains correct values

* Lots of changes Read is working correctly

* Bug fixes

* Pep 8 fixes

* comments

* merge with master

* A few comment updates

* Removing string based implementation

Now both read and write will calculate the read and write values using
the calculated delta.  This as a response to the review. In doing so I
was able to remove the fixedPointHelper, the only thing left in that
file was the warn, which I moved to a method of session._FXPRegister

I updated the tests to now work with the correct implementation my next
commit will include a test clean up

* Unit test clean -up

* Refactor overflow bit

Rather than its own seperate property overflow status is now read as a
tuple and written to as a tuple for a FXP register with overflow
enabled.
Unit tests again all post, required some refactoring

* Increase precision to not have rounding for larger integer words

* Bitfile changes

* Fix bug in reading 64 bit FXP with overflow

automated hardware tests now all work.

* Make FxpRegister inherit from Register not ArrayRegister

* Did not mean to commit these

* Comments

* Flake 8 Fixes

removing check for w503 that flake 8 seems to think
foo(5 + 6 +
7 + 8)
is better than
foo(5 + 6
+ 7 + 8)
Even though Pep 8 says otherwise

* Pep 8

* Fix 3.x builds

* 3.x Does not support error.message

Removing the custom input text. Users must look at documentation now

* Responding to reviews

* Small fixes for reviews

* Remove new line

* Spelling